+++ src/main/java/com/breakinblocks/neosync/common/block/SyncBlocks.java
@@ -57,6 +57,19 @@ public class SyncBlocks {
                     .isSuffocating(SyncBlocks::never)
                     .isViewBlocking(SyncBlocks::never)));
 
+    public static final DeferredBlock<ManualShellStorageBlock> MANUAL_SHELL_STORAGE = BLOCKS.register("manual_shell_storage",
+            name -> new ManualShellStorageBlock(BlockBehaviour.Properties.ofFullCopy(Blocks.GLASS)
+                    .setId(ResourceKey.create(Registries.BLOCK, name))
+                    .mapColor(MapColor.COLOR_GRAY)
+                    .requiresCorrectToolForDrops()
+                    .strength(1.8F)
+                    .sound(SoundType.GLASS)
+                    .noOcclusion()
+                    .isValidSpawn(SyncBlocks::never)
+                    .isRedstoneConductor(SyncBlocks::never)
+                    .isSuffocating(SyncBlocks::never)
+                    .isViewBlocking(SyncBlocks::never)));
+
     public static final DeferredBlock<ZeroPointShellConstructorBlock> ZERO_POINT_SHELL_CONSTRUCTOR = BLOCKS.register("zero_point_shell_constructor",
             name -> new ZeroPointShellConstructorBlock(BlockBehaviour.Properties.ofFullCopy(Blocks.GLASS)
                     .setId(ResourceKey.create(Registries.BLOCK, name))